+++ air-helper.php
@@ -156,11 +156,19 @@ add_filter( 'bulk_actions-plugins','air_helper_modify_plugins_bulk_actions' );
 // @codingStandardsIgnoreEnd
 
 /**
+ *  Require priority files.
+ */
+function air_helper_priority_fly() {
+	load_plugin_textdomain( 'air-helper', false, dirname( plugin_basename( __FILE__ ) ) . '/languages/' );
+
+	require_once air_helper_base_path() . '/inc/hooks-priority.php';
+}
+add_action( 'init', 'air_helper_priority_fly', 5 );
+
+/**
  *  Require files containing our preferences.
  */
 function air_helper_fly() {
-	load_plugin_textdomain( 'air-helper', false, dirname( plugin_basename( __FILE__ ) ) . '/languages/' );
-
 	require_once air_helper_base_path() . '/inc/hooks.php';
 	require_once air_helper_base_path() . '/inc/functions.php';
 	require_once air_helper_base_path() . '/inc/misc.php';