+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
+### Added
+- Priority hooks file loaded in `init` hook with priority `5`
+- User enumeration stop
 
 ## [1.7.3] - 2019-01-08
 ### Fixed